requirements-all.txt
Flask>=2.3.0
PyYAML>=6.0.0
accelerate>=0.21.0
datasets>=2.13.0
langchain-core>=0.0.1
langchain-text-splitters>=0.0.1
milvus-lite>=2.4.0
modelscope>=1.19.0
neo4j>=5.0.0
numpy>=1.24.0
openpyxl>=3.1.0
pdfplumber>=0.7.0
peft>=0.4.0
protobuf>=3.20.0
pymilvus>=2.0.0
pymysql>=1.0.0
python-docx>=0.8.11
requests>=2.31.0
scipy>=1.10.0
sentence-transformers>=2.0.0
sentencepiece>=0.1.99
torch>=2.0.0
tqdm>=4.65.0
transformers>=4.30.0
trl>=0.9.0
xlrd>=2.0.1
bitsandbytes>=0.43.0; platform_system == "Linux"
